# swagger-spec.yml
openapi: 3.0.0
paths:
  /api:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        "200":
          description: ""
  /api/login:
    post:
      operationId: AppController_login
      parameters: []
      responses:
        "201":
          description: ""
  /api/profile:
    get:
      operationId: AppController_getProfile
      parameters: []
      responses:
        "200":
          description: ""
  /api/guardians:
    get:
      operationId: GuardianController_getAllGuardians
      parameters: []
      responses:
        "200":
          description: ""
    post:
      operationId: GuardianController_addGuardian
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuardianDto"
      responses:
        "201":
          description: ""
  "/api/guardians/{id}":
    get:
      operationId: GuardianController_getGuardianById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
info:
  title: guardian-service
  description: Resource Server
  version: 0.1.0.alpha
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateGuardianDto:
      type: object
      properties:
        registrationNumber:
          type: string
        displayName:
          type: string
        description:
          type: string
        wallet:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
        email:
          type: string
        phonenumber:
          type: string
        website:
          type: string
        identityCommitment:
          type: string
      required:
        - registrationNumber
        - displayName
        - description
        - wallet
        - address
        - city
        - state
        - zip
        - country
        - email
        - phonenumber
        - website
        - identityCommitment